Author: therve
Reviewers: exarkun, jml, cablehead
Fixes #2855

Add tests to the twisted.protocols.ftp.IFTPShell interface, to test the
implementation in ftp itself, and to allow to test against other implementations
in the future (eg, vfs).


git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@21592 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/protocols/ftp.py
@@ -95,6 +95,8 @@ PERMISSION_DENIED                       = "550.2"     # permission denied
 ANON_USER_DENIED                        = "550.3"     # anonymous users can't alter filesystem
 IS_NOT_A_DIR                            = "550.4"     # rmd called on a path that is not a directory
 REQ_ACTN_NOT_TAKEN                      = "550.5"
+FILE_EXISTS                             = "550.6"
+IS_A_DIR                                = "550.7"
 PAGE_TYPE_UNK                           = "551"
 EXCEEDED_STORAGE_ALLOC                  = "552"
 FILENAME_NOT_ALLOWED                    = "553"
@@ -160,14 +162,20 @@ RESPONSE = {
     PERMISSION_DENIED:                  '550 %s: Permission denied.',
     ANON_USER_DENIED:                   '550 Anonymous users are forbidden to change the filesystem',
     IS_NOT_A_DIR:                       '550 Cannot rmd, %s is not a directory',
+    FILE_EXISTS:                        '550 %s: File exists',
+    IS_A_DIR:                           '550 %s: is a directory',
     REQ_ACTN_NOT_TAKEN:                 '550 Requested action not taken: %s',
     EXCEEDED_STORAGE_ALLOC:             '552 Requested file action aborted, exceeded file storage allocation',
     FILENAME_NOT_ALLOWED:               '553 Requested action not taken, file name not allowed'
 }
 
 
+
 class InvalidPath(Exception):
-    pass
+    """
+    Internal exception used to signify an error during parsing a path.
+    """
+
 
 
 def toSegments(cwd, path):
@@ -196,44 +204,47 @@ def toSegments(cwd, path):
 
 
 def errnoToFailure(e, path):
-    if e == errno.ENOENT or e == errno.EISDIR:
+    """
+    Map C{OSError} and C{IOError} to standard FTP errors.
+    """
+    if e == errno.ENOENT:
         return defer.fail(FileNotFoundError(path))
     elif e == errno.EACCES or e == errno.EPERM:
         return defer.fail(PermissionDeniedError(path))
     elif e == errno.ENOTDIR:
         return defer.fail(IsNotADirectoryError(path))
+    elif e == errno.EEXIST:
+        return defer.fail(FileExistsError(path))
+    elif e == errno.EISDIR:
+        return defer.fail(IsADirectoryError(path))
     else:
         return defer.fail()
 
-# Generic VFS exceptions
-class FilesystemShellException(Exception):
-    """Base for IFTPShell errors.
-    """
 
-class NoSuchFile(FilesystemShellException):
-    pass
-
-class PermissionDenied(FilesystemShellException):
-    pass
-
-class WrongFiletype(FilesystemShellException):
-    pass
-
-
-# -- Custom Exceptions --
 
 class FTPCmdError(Exception):
+    """
+    Generic exception for FTP commands.
+    """
     def __init__(self, *msg):
         Exception.__init__(self, *msg)
         self.errorMessage = msg
 
+
     def response(self):
-        """Generate a FTP response message for this error."""
+        """
+        Generate a FTP response message for this error.
+        """
         return RESPONSE[self.errorCode] % self.errorMessage
 
 
+
 class FileNotFoundError(FTPCmdError):
+    """
+    Raised when trying to access a non existent file or directory.
+    """
     errorCode = FILE_NOT_FOUND
+
 
 
 class AnonUserDeniedError(FTPCmdError):
@@ -248,12 +259,14 @@ class AnonUserDeniedError(FTPCmdError):
     errorCode = ANON_USER_DENIED
 
 
+
 class PermissionDeniedError(FTPCmdError):
     """
     Raised when access is attempted to a resource to which access is
     not allowed.
     """
     errorCode = PERMISSION_DENIED
+
 
 
 class IsNotADirectoryError(FTPCmdError):
@@ -263,12 +276,38 @@ class IsNotADirectoryError(FTPCmdError):
     errorCode = IS_NOT_A_DIR
 
 
+
+class FileExistsError(FTPCmdError):
+    """
+    Raised when attempted to override an existing resource.
+    """
+    errorCode = FILE_EXISTS
+
+
+
+class IsADirectoryError(FTPCmdError):
+    """
+    Raised when DELE is called on a path that is a directory.
+    """
+    errorCode = IS_A_DIR
+
+
+
 class CmdSyntaxError(FTPCmdError):
+    """
+    Raised when a command syntax is wrong.
+    """
     errorCode = SYNTAX_ERR
 
 
+
 class CmdArgSyntaxError(FTPCmdError):
+    """
+    Raised when a command is called with wrong value or a wrong number of
+    arguments.
+    """
     errorCode = SYNTAX_ERR_IN_ARGS
+
 
 
 class CmdNotImplementedError(FTPCmdError):
@@ -278,44 +317,41 @@ class CmdNotImplementedError(FTPCmdError):
     errorCode = CMD_NOT_IMPLMNTD
 
 
+
 class CmdNotImplementedForArgError(FTPCmdError):
+    """
+    Raised when the handling of a parameter for a command is not implemented by
+    the server.
+    """
     errorCode = CMD_NOT_IMPLMNTD_FOR_PARAM
+
+
 
 class FTPError(Exception):
     pass
 
-class FTPTimeoutError(Exception):
-    pass
 
-class DTPError(Exception):
-    pass
-
-class BogusClientError(Exception):
-    """thrown when a client other than the one we opened this
-    DTP connection for attempts to connect, or a client attempts to
-    get us to connect to an ip that differs from the one where the
-    request came from"""
-    pass
-
-class PathBelowTLDError(FTPCmdError):
-    errorCode = PERMISSION_DENIED
-
-class ClientDisconnectError(Exception):
-    pass
 
 class PortConnectionError(Exception):
     pass
 
+
+
 class BadCmdSequenceError(FTPCmdError):
-    """raised when a client sends a series of commands in an illogical sequence"""
+    """
+    Raised when a client sends a series of commands in an illogical sequence.
+    """
     errorCode = BAD_CMD_SEQ
 
+
+
 class AuthorizationError(FTPCmdError):
-    """raised when client authentication fails"""
+    """
+    Raised when client authentication fails.
+    """
     errorCode = AUTH_FAILURE
 
 
-# -- Utility Functions --
 
 def debugDeferred(self, *_):
     log.msg('debugDeferred(): %s' % str(_), debug=True)
@@ -328,6 +364,7 @@ _months = [
     None,
     'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun',
     'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec']
+
 
 class DTP(object, protocol.Protocol):
     implements(interfaces.IConsumer)
@@ -1216,6 +1253,7 @@ class FTPFactory(policies.LimitTotalConnectionsFactory):
 
 # -- Cred Objects --
 
+
 class IFTPShell(Interface):
     """
     An abstraction of the shell commands used by the FTP protocol for
@@ -1352,9 +1390,12 @@ class IFTPShell(Interface):
         """
 
 
+
 class IReadFile(Interface):
-    """A file out of which bytes may be read.
     """
+    A file out of which bytes may be read.
+    """
+
     def send(consumer):
         """
         Produce the contents of the given path to the given consumer.  This
@@ -1367,9 +1408,12 @@ class IReadFile(Interface):
         """
 
 
+
 class IWriteFile(Interface):
-    """A file into which bytes may be written.
     """
+    A file into which bytes may be written.
+    """
+
     def receive():
         """
         Create a consumer which will write to this file.  This method may
@@ -1377,6 +1421,7 @@ class IWriteFile(Interface):
 
         @rtype: C{Deferred} of C{IConsumer}
         """
+
 
 
 def _getgroups(uid):
@@ -1450,6 +1495,7 @@ def _testPermissions(uid, gid, spath, mode='r'):
     return access
 
 
+
 class FTPAnonymousShell(object):
     """
     An anonymous implementation of IFTPShell
@@ -1463,27 +1509,38 @@ class FTPAnonymousShell(object):
     def __init__(self, filesystemRoot):
         self.filesystemRoot = filesystemRoot
 
+
     def _path(self, path):
         return reduce(filepath.FilePath.child, path, self.filesystemRoot)
+
 
     def makeDirectory(self, path):
         return defer.fail(AnonUserDeniedError())
 
+
     def removeDirectory(self, path):
         return defer.fail(AnonUserDeniedError())
+
 
     def removeFile(self, path):
         return defer.fail(AnonUserDeniedError())
 
+
     def rename(self, fromPath, toPath):
         return defer.fail(AnonUserDeniedError())
+
 
     def receive(self, path):
         path = self._path(path)
         return defer.fail(AnonUserDeniedError())
 
+
     def openForReading(self, path):
         p = self._path(path)
+        if p.isdir():
+            # Normally, we would only check for EISDIR in open, but win32
+            # returns EACCES in this case, so we check before
+            return defer.fail(IsADirectoryError(path))
         try:
             f = p.open('rb')
         except (IOError, OSError), e:
@@ -1514,46 +1571,78 @@ class FTPAnonymousShell(object):
         else:
             return defer.succeed(None)
 
+
     def stat(self, path, keys=()):
         p = self._path(path)
         if p.isdir():
-            return defer.fail(WrongFiletype())
+            try:
+                statResult = self._statNode(p, keys)
+            except (IOError, OSError), e:
+                return errnoToFailure(e.errno, path)
+            except:
+                return defer.fail()
+            else:
+                return defer.succeed(statResult)
         else:
             return self.list(path, keys).addCallback(lambda res: res[0][1])
 
+
     def list(self, path, keys=()):
-        path = self._path(path)
-        if path.isdir():
-            entries = path.listdir()
+        """
+        Return the list of files at given C{path}, adding C{keys} stat
+        informations if specified.
+
+        @param path: the directory or file to check.
+        @type path: C{str}
+
+        @param keys: the list of desired metadata
+        @type keys: C{list} of C{str}
+        """
+        filePath = self._path(path)
+        if filePath.isdir():
+            entries = filePath.listdir()
+            fileEntries = [filePath.child(p) for p in entries]
+        elif filePath.isfile():
+            entries = [os.path.join(*filePath.segmentsFrom(self.filesystemRoot))]
+            fileEntries = [filePath]
         else:
-            entries = [None]
+            return defer.fail(FileNotFoundError(path))
 
         results = []
-        for fName in entries:
+        for fileName, filePath in zip(entries, fileEntries):
             ent = []
-            results.append((fName, ent))
+            results.append((fileName, ent))
             if keys:
-                if fName is not None:
-                    p = os.path.join(path.path, fName)
-                else:
-                    p = path.path
                 try:
-                    statResult = os.stat(p)
+                    ent.extend(self._statNode(filePath, keys))
                 except (IOError, OSError), e:
-                    return errnoToFailure(e.errno, path)
+                    return errnoToFailure(e.errno, fileName)
                 except:
                     return defer.fail()
 
-            for k in keys:
-                ent.append(getattr(self, '_list_' + k)(statResult))
         return defer.succeed(results)
 
-    _list_size = operator.attrgetter('st_size')
-    _list_permissions = operator.attrgetter('st_mode')
-    _list_hardlinks = operator.attrgetter('st_nlink')
-    _list_modified = operator.attrgetter('st_mtime')
 
-    def _list_owner(self, st):
+    def _statNode(self, filePath, keys):
+        """
+        Shortcut method to get stat info on a node.
+
+        @param filePath: the node to stat.
+        @type filePath: C{filepath.FilePath}
+
+        @param keys: the stat keys to get.
+        @type keys: C{iterable}
+        """
+        filePath.restat()
+        return [getattr(self, '_stat_' + k)(filePath.statinfo) for k in keys]
+
+    _stat_size = operator.attrgetter('st_size')
+    _stat_permissions = operator.attrgetter('st_mode')
+    _stat_hardlinks = operator.attrgetter('st_nlink')
+    _stat_modified = operator.attrgetter('st_mtime')
+
+
+    def _stat_owner(self, st):
         if pwd is not None:
             try:
                 return pwd.getpwuid(st.st_uid)[0]
@@ -1561,7 +1650,8 @@ class FTPAnonymousShell(object):
                 pass
         return str(st.st_uid)
 
-    def _list_group(self, st):
+
+    def _stat_group(self, st):
         if grp is not None:
             try:
                 return grp.getgrgid(st.st_gid)[0]
@@ -1569,8 +1659,10 @@ class FTPAnonymousShell(object):
                 pass
         return str(st.st_gid)
 
-    def _list_directory(self, st):
+
+    def _stat_directory(self, st):
         return bool(st.st_mode & stat.S_IFDIR)
+
 
 
 class _FileReader(object):
@@ -1593,11 +1685,18 @@ class _FileReader(object):
         return d
 
 
+
 class FTPShell(FTPAnonymousShell):
+    """
+    An authenticated implementation of L{IFTPShell}.
+    """
+
     def makeDirectory(self, path):
         p = self._path(path)
         try:
             p.makedirs()
+        except (IOError, OSError), e:
+            return errnoToFailure(e.errno, path)
         except:
             return defer.fail()
         else:
@@ -1606,8 +1705,15 @@ class FTPShell(FTPAnonymousShell):
 
     def removeDirectory(self, path):
         p = self._path(path)
+        if p.isfile():
+            # Win32 returns the wrong errno when rmdir is called on a file
+            # instead of a directory, so as we have the info here, let's fail
+            # early with a pertinent error
+            return defer.fail(IsNotADirectoryError(path))
         try:
             os.rmdir(p.path)
+        except (IOError, OSError), e:
+            return errnoToFailure(e.errno, path)
         except:
             return defer.fail()
         else:
@@ -1616,8 +1722,15 @@ class FTPShell(FTPAnonymousShell):
 
     def removeFile(self, path):
         p = self._path(path)
+        if p.isdir():
+            # Win32 returns the wrong errno when remove is called on a
+            # directory instead of a file, so as we have the info here,
+            # let's fail early with a pertinent error
+            return defer.fail(IsADirectoryError(path))
         try:
             p.remove()
+        except (IOError, OSError), e:
+            return errnoToFailure(e.errno, path)
         except:
             return defer.fail()
         else:
@@ -1629,6 +1742,8 @@ class FTPShell(FTPAnonymousShell):
         tp = self._path(toPath)
         try:
             os.rename(fp.path, tp.path)
+        except (IOError, OSError), e:
+            return errnoToFailure(e.errno, fromPath)
         except:
             return defer.fail()
         else:
@@ -1637,11 +1752,18 @@ class FTPShell(FTPAnonymousShell):
 
     def openForWriting(self, path):
         p = self._path(path)
+        if p.isdir():
+            # Normally, we would only check for EISDIR in open, but win32
+            # returns EACCES in this case, so we check before
+            return defer.fail(IsADirectoryError(path))
         try:
             fObj = p.open('wb')
+        except (IOError, OSError), e:
+            return errnoToFailure(e.errno, path)
         except:
             return defer.fail()
         return defer.succeed(_FileWriter(fObj))
+
 
 
 class _FileWriter(object):
@@ -1656,6 +1778,7 @@ class _FileWriter(object):
         self._receive = True
         # FileConsumer will close the file object
         return defer.succeed(FileConsumer(self.fObj))
+
 
 
 class FTPRealm:
